requirements.txt
torch
numpy
matplotlib
pyyaml
tqdm
flake8
pytest
pytest-cov
pandas